+++ lib/huevo.dart
@@ -1,5 +1,5 @@
 library huevo;
 
-export 'src/parser.dart' show Parser, StringParser;
+export 'src/parser.dart' show Parser, StringParser, RegexParser;
 export 'src/predef.dart';
 export 'src/input_source.dart';