+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0-DEV
 
 - Initial version
+- Provide StringParser, and RegexParser